backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-dotenv==1.0.1
redis==5.0.8
httpx==0.27.2
anthropic==0.34.2
openai==1.51.0
chromadb==0.5.11
numpy==1.26.4
python-dateutil==2.9.0.post0
loguru==0.7.2
pytest==8.3.3
pytest-asyncio==0.24.0
